refactor: NewLedgerFromSegmentsBuilder を抽出 (Issue #1284)
